requirements.txt
pandas>=2.0
numpy>=1.24
beautifulsoup4>=4.12
lxml>=4.9
requests>=2.28
google-api-python-client>=2.100
google-auth>=2.22
google-auth-oauthlib>=1.1
streamlit>=1.35
